Create activities package. Move activities to package.
