+++ app/src/main/res/layout/activity_add_weather_item.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".AddWeatherItemActivity">
+        tools:context=".activities.AddWeatherItemActivity">
 
         <TextView
             android:layout_width="wrap_content"
