+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity" >
+    tools:context=".activities.MainActivity" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerview"
